Adding the parameters

+++ vars/regressionTests.groovy
@@ -1,14 +1,16 @@
-def executeRegressionTests(browser) {
+def executeRegressionTests(heal_enabled, levels, browser) {
     print('Executing Regression tests')
-    print(browser)
+    print(heal_enabled, levels, browser)
     node {
         stage('Checkout Tests') {
+            if (fileExists('project')) {
+                sh 'rm -rf project'
+            }
             dir('project') {
                 git branch: 'master', credentialsId: 'bitbucket', url: 'https://bitbucket.org/haisrig/sfd_app_rtcs.git'
                 sh 'ls -ltR'
                 print('Cloing the Regression tests from Bitbucket')
-            }
-            sh 'rm -rf project'
+            }            
         }
         stage('Prepare Environment') {
             print('Preparing Environment for tests')
